% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kevinsRepo-package.R
\docType{package}
\name{kevinsRepo-package}
\alias{kevinsRepo}
\alias{kevinsRepo-package}
\title{kevinsRepo: `kevinsRepo` Is A Personal Utility, Data, and Testing Package}
\description{
This package has three purposes: 1) to store utility functions I use frequently, 2) to store my data, and 3) to provide a testing ground for new functions I develop and utilize. Please note that the functions and data in this package are not guaranteed to be bug-free or optimized. This package is for personal use only, and everything in it should be assumed to be under development.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/knavarrete-noble/kevinsRepo}
  \item Report bugs at \url{https://github.com/knavarrete-noble/kevinsRepo/issues}
}

}
\author{
\strong{Maintainer}: Kevin Navarrete-Parra \email{navark4@unlv.nevada.edu}

}
\keyword{internal}
