% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_diag.R
\name{logit_diag}
\alias{logit_diag}
\title{Common Logistic Regression Diagnostic Checks}
\usage{
logit_diag(model, new_data)
}
\arguments{
\item{model}{A logit model of class glm}

\item{model_name}{character string for the prefix name that will accompany all objects generated}

\item{counter}{Integer corresponding to the model being evaluated}
}
\value{
A series of diagnostic checks for a logistic regression model exported to the global environment
}
\description{
Common Logistic Regression Diagnostic Checks
}
\examples{
data <- data.frame(y = c(1, 0, 1, 0, 1, 0, 1, 0, 1, 0),
                   x1 = c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10),
                   x2 = c(10, 9, 8, 7, 6, 5, 4, 3, 2, 1))

model <- glm(y ~ x1 + x2, data = data, family = binomial)

new_data <- data.frame(x1 = c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10),
                      x2 = c(10, 9, 8, 7, 6, 5, 4, 3, 2, 1),
                      y = c(1, 0, 1, 0, 1, 0, 1, 0, 1, 0))

logit_diag(model = model, new_data = new_data)
}
