% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bingo.R
\name{make_bingo}
\alias{make_bingo}
\title{Make a bingo card}
\usage{
make_bingo(
  str,
  wrap = 20,
  n_cards = 5,
  free_space = "Free",
  title = "Bingo!",
  file_type = ".pdf",
  save = FALSE
)
}
\arguments{
\item{str}{A character vector of at least length 25 denoting the contents of the bingo card}

\item{wrap}{A numeric value indicating the number of characters to wrap the text at}

\item{n_cards}{A numeric value indicating the number of bingo cards to create}

\item{free_space}{A character string to customize what the free space says}

\item{title}{A character string to customize the title of the bingo card}

\item{file_type}{A character string indicating the file type to save the bingo card as. Default is '.pdf', but can also be '.png', '.jpeg', or '.svg', among others supported by the \code{ggplot2::ggsave} function.}

\item{save}{A logical value indicating whether or not to save the bingo cards. Default is FALSE.}
}
\value{
A ggplot object of the bingo card(s) with the specified contents
}
\description{
Make a bingo card
}
\details{
This function repurposes code found in Quang Nguyen's blog post on creating bingo cards using the ggplot2 package. The original blog post can be found at https://qntkhvn.netlify.app/projects/bingo. I've updated the original code to include additional parameters as well as the additional string wrapping functionality to ensure that the text fits within the bingo card cells.
}
\examples{
# Create a bingo card with the default parameters
make_bingo(letters)
}
