% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_acc.R
\name{tab_acc}
\alias{tab_acc}
\title{Confusion Matrix}
\usage{
tab_acc(class.obj)
}
\arguments{
\item{class.obj}{A confusion matrix of class table}
}
\value{
A character string of the accuracy of the confusion matrix
}
\description{
Confusion Matrix
}
\examples{
# Example data
data <- data.frame(y = c(1, 0, 1, 0, 1, 0, 1, 0, 1, 0),
                  x = c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10))

# Fit a logistic regression model
model <- glm(y ~ x, data = data, family = binomial)

# Predict the model
data$pred <- ifelse(predict(model, data, type = "response") > 0.5, 1, 0)

# Create a confusion matrix
class.obj <- table(data$y, data$pred)

# Calculate the accuracy of the confusion matrix
tab_acc(class.obj)
}
